# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Experiment Infrastructure",
    "version": "0.2.0"
  },
  "paths": {
...
    "/experiments": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Create",
        "operationId": "create_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperiment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Experiment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}": {
      "get": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Get",
        "operationId": "get_experiments__experiment_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Experiment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}/state": {
      "get": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "State",
        "operationId": "state_experiments__experiment_id__state_get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/experiment_infra__runtime__api__Result"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}/hypotheses": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Hypothesis",
        "operationId": "hypothesis_experiments__experiment_id__hypotheses_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HypothesisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hypothesis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}/runs": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Run",
        "operationId": "run_experiments__experiment_id__runs_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}/actions": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Action",
        "operationId": "action_experiments__experiment_id__actions_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryStep"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}/claims": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Claim",
        "operationId": "claim_experiments__experiment_id__claims_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/experiment_infra__runtime__api__ClaimRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/experiment_infra__core__schema__Claim"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}/verify": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Verify",
        "operationId": "verify_experiments__experiment_id__verify_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/experiment_infra__runtime__api__VerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/replay": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Replay",
        "operationId": "replay_runs__run_id__replay_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/branch": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Branch",
        "operationId": "branch_runs__run_id__branch_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/experiment_infra__runtime__api__BranchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/finish": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Finish Run",
        "operationId": "finish_run_runs__run_id__finish_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{experiment_id}/finish": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Finish Experiment",
        "operationId": "finish_experiment_experiments__experiment_id__finish_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Experiment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{object_id}/provenance": {
      "get": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Provenance",
        "operationId": "provenance_objects__object_id__provenance_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/experiment_infra__runtime__api__Result"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trajectories/{trajectory_id}": {
      "get": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Trajectory",
        "operationId": "trajectory_trajectories__trajectory_id__get",
        "parameters": [
          {
            "name": "trajectory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trajectory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trajectory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exports": {
      "post": {
        "tags": [
          "Experiment infrastructure"
        ],
        "summary": "Export",
        "operationId": "export_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/experiment_infra__runtime__api__Result"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActionRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "action": {
            "$ref": "#/components/schemas/Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "action"
        ],
        "title": "ActionRequest"
      },
...
      "CreateExperiment": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "goal": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Goal"
          },
          "environment_spec": {
            "$ref": "#/components/schemas/EnvironmentSpec"
          },
          "agent_backend": {
            "type": "string",
            "title": "Agent Backend",
            "default": "external"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "success_criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Success Criteria"
          },
          "budget": {
            "additionalProperties": true,
            "type": "object",
            "title": "Budget"
          },
          "parent_experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Experiment Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "CreateExperiment"
      },
...
      "Experiment": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "goal": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Goal"
          },
          "hypothesis_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hypothesis Ids"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ],
            "title": "Status",
            "default": "CREATED"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "success_criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Success Criteria"
          },
          "budget": {
            "additionalProperties": true,
            "type": "object",
            "title": "Budget"
          },
          "environment_spec": {
            "$ref": "#/components/schemas/EnvironmentSpec"
          },
          "agent_backend": {
            "type": "string",
            "title": "Agent Backend",
            "default": "external"
          },
          "parent_experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Experiment Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "Experiment"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hypothesis": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "hypothesis_id": {
            "type": "string",
            "title": "Hypothesis Id"
          },
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "statement": {
            "type": "string",
            "minLength": 1,
            "title": "Statement"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "default": ""
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "expected_observation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Expected Observation"
          },
          "status": {
            "type": "string",
            "enum": [
              "PROPOSED",
              "TESTED"
            ],
            "title": "Status",
            "default": "PROPOSED"
          },
          "parent_hypothesis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Hypothesis Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "experiment_id",
          "statement"
        ],
        "title": "Hypothesis"
      },
      "HypothesisRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "default": ""
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "expected_observation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Expected Observation"
          },
          "parent_hypothesis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Hypothesis Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "statement"
        ],
        "title": "HypothesisRequest"
      },
...
      "ReplayResult": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "original_run_id": {
            "type": "string",
            "title": "Original Run Id"
          },
          "replay_run_id": {
            "type": "string",
            "title": "Replay Run Id"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched"
          },
          "comparisons": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Comparisons"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "original_run_id",
          "replay_run_id",
          "matched",
          "comparisons"
        ],
        "title": "ReplayResult"
      },
...
      "RunRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "hypothesis_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis Ids"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunRequest"
      },
...
      "experiment_infra__runtime__api__ClaimRequest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "statement"
        ],
        "title": "ClaimRequest"
      },
      "experiment_infra__runtime__api__Result": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "2.0",
            "title": "Schema Version",
            "default": "2.0"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Result"
      },
...
    }
  }
}